+++ packages/espressocash_app/lib/l10n/intl_en.arb
@@ -1265,5 +1265,7 @@
     "changeDocumentPhoto": "Change",
     "@changeDocumentPhoto": {},
     "selectDocumentType": "Select Document Type",
-    "@selectDocumentType": {}
+    "@selectDocumentType": {},
+    "documentTitle": "Document",
+    "@documentTitle": {}
 }
